Move console functionality into pal_console
Just moving some code around.
