+++ src/Native/System.Native/pal_io.h
@@ -22,17 +22,6 @@ struct FileStatus
     int64_t BirthTime; // time the file was created
 };
 
-/*
- * Window Size of the terminal
-*/
-struct WinSize
-{
-    uint16_t Row;
-    uint16_t Col;
-    uint16_t XPixel;
-    uint16_t YPixel;
-};
-
 /************
  * The values below in the header are fixed and correct for managed callers to use forever.
  * We must never change them. The implementation must either static_assert that they are equal
@@ -641,28 +630,6 @@ extern "C" void Sync();
  * Returns the number of bytes written on success; otherwise, returns -1 and sets errno
  */
 extern "C" int32_t Write(int32_t fd, const void* buffer, int32_t bufferSize);
-
-/**
- * Gets the windows size of the terminal
- *
- * Returns 0 on success; otherwise, returns errorNo.
- */
-extern "C" int32_t GetWindowSize(WinSize* windowsSize);
-
-/**
- * Gets whether the specified file descriptor is for a terminal.
- *
- * Returns 1 if the file descriptor is referring to a terminal;
- * otherwise returns 0 and sets errno.
- */
-extern "C" int32_t IsATty(int filedes);
-
-/**
-* Reads the number of bytes specified into the provided buffer from stdin.
-* in a non-echo and non-canonical mode.
-* Returns the number of bytes read on success; otherwise, -1 is returned an errno is set.
-*/
-extern "C" int32_t ReadStdinUnbuffered(void* buffer, int32_t bufferSize);
 
 /**
 * Initializes a new inotify instance and returns a file
